+++ src/main/java/com/example/gdg3/search/JpaSearchRepository.java
@@ -1,4 +1,4 @@
-package search;
+package com.example.gdg3.search;
 
 public class JpaSearchRepository {
 }
