+++ src/main/java/com/example/gdg3/search/SearchController.java
@@ -1,4 +1,4 @@
-package search;
+package com.example.gdg3.search;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
