+++ src/main/java/com/example/gdg3/search/SearchService.java
@@ -1,4 +1,4 @@
-package search;
+package com.example.gdg3.search;
 
 public interface SearchService {
     // public search(String location, int pageNum, int size);
